refactor: extract useEditorTabSwap hook from Editor — reduce complexity
